backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
boto3==1.35.99
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.20
moviepy
Pillow
certifi
google-genai
